requirements.txt
tabula.py
pandas
numpy
beautifulsoup4
xlrd